+++ riscv/src/gui_egui/components/rv_mem.rs
@@ -38,7 +38,6 @@ impl RVMem {
                         ((self.range.end - self.range.start) / 4) as usize,
                         |mut row| {
                             let index = row.index();
-                            //println!("{}", index);
                             let address = self.range.start as usize + index * 4;
                             let memory = self.memory.0.borrow().clone();
                             row.col(|ui| {