+++ src/gui_egui/gui.rs
@@ -26,6 +26,7 @@ pub struct Gui {
     pub clip_rect: Rect,
     pub shortcuts: Shortcuts,
     pub pause: bool,
+    pub step_amount: usize, //TODO change this to be a menu struct, and maybe move pause and other here
     pub editor: Option<Editor>,
     pub editor_use: bool,
     pub contexts: HashMap<crate::common::Id, EguiExtra>,
@@ -57,6 +58,7 @@ pub fn gui(cs: ComponentStore, path: &PathBuf, library: Library) -> Result<(), e
         clip_rect: Rect::NOTHING,
         shortcuts: Shortcuts::new(),
         pause: true,
+        step_amount: 10,
         editor: None,
         editor_use: false,
         contexts,
@@ -103,7 +105,7 @@ impl eframe::App for Gui {
             self.top_bar(ctx);
             if self.simulator.is_some() {
                 // self.side_panel(ctx);
-                if self.simulator.as_ref().unwrap().running {
+                if self.simulator.as_ref().unwrap().is_running() {
                     self.simulator.as_mut().unwrap().run();
 
                     // This makes the ui run agin as to not stop the simulation