+++ src/gui_egui/menu.rs
@@ -35,8 +35,36 @@ impl Menu {
             if ui.button("⏸").clicked() {
                 keymap::control_pause_fn(gui);
             }
+            ui.separator();
+
+            ui.add(DragValue::new(&mut gui.step_amount).prefix("Step: "));
+            if ui.button("⟳").clicked() {
+                // TODO dont have simulator here add keymap
+                if let Some(s) = gui.simulator.as_mut() {
+                    let _ = s.set_step_to(s.cycle + gui.step_amount);
+                }
+            }
+
+            ui.separator();
+
             if let Some(s) = gui.simulator.as_ref() {
-                ui.label(format!("Cycle #{}", s.cycle));
+                ui.label(format!("Cycle: {}", s.cycle));
+
+                // This displays the current state of the simulation,
+                // if hovered displays the component condition vector
+                let r = ui.label(format!("State: {:?}", s.get_state()));
+
+                // if there are some components  that reported a condition show it when
+                // state is hovered over
+                if let Some(cond_vec) = s.get_component_condition() {
+                    r.on_hover_text(
+                        cond_vec // format all id condition pairs and sum them to a string
+                            .iter()
+                            .fold(String::from("Reported conditions"), |acc, id_cond| {
+                                acc + &format!("\nID: {}, {:?}", id_cond.0, id_cond.1)
+                            }),
+                    );
+                }
             }
         });
     }