+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 Tracking changes per date:
 
+## 240911
+- Added so that simulator now tracks what component condition
+- Added simulator running state, such as running, halt or error
+- Added stepping functionality
+- Added functionality to gui to show running state and component condition
+
+## 240909
+
+- Added un-clock history for set of active components
+
+## 240908
+
+- Added functionality to determine set of active components
+
 ## 230801
 
 - Simulator run/halt is implemented in `vizia` using a simple eventing mechanism. Later we might want to spawn a simulation thread for faster execution (right now its tied to frame rate).